+++ tests/testthat/test-emoji.R
@@ -1,0 +1,11 @@
+library(testthat)
+library(emoji)
+
+test_that("Both aliases work with emoji()", {
+  aliases <- emojis$aliases[[1]]
+
+  expect_identical(
+    emoji(aliases[1]),
+    emoji(aliases[2])
+  )
+})