+++ NEWS.md
@@ -1,6 +1,7 @@
 # emoji (development version)
 
 * `emoji_name` now includes all aliases.
+* Adding `emoji()` and `emoji_find()` which can be used to return emojis based on keywords.
 
 # emoji 0.1.0